requirements.txt
exa_py
python-dotenv
beautifulsoup4
lxml
lxml_html_clean
openai
trafilatura
python-docx
flask
gunicorn
